+++ src/budge/account.py
@@ -4,6 +4,7 @@ from heapq import merge
 from itertools import groupby
 from typing import Generator, NamedTuple
 
+from dateutil.relativedelta import relativedelta
 from stockholm import Money
 
 from .collection import Collection
@@ -11,7 +12,12 @@ from .date import daterange
 from .transaction import RepeatingTransaction, Transaction
 
 
-class Balance(NamedTuple):
+class RunningBalance(NamedTuple):
+    transaction: Transaction
+    balance: Money
+
+
+class DailyBalance(NamedTuple):
     date: date
     balance: Money
 
@@ -60,6 +66,21 @@ class Account:
                 break
             yield transaction
 
+    def running_balance(
+        self, start_date: date | None = None, end_date: date | None = None
+    ) -> Generator[RunningBalance]:
+        """Iterate over transactions in the account over the given range with a
+        running account balance."""
+        balance = (
+            self.balance(start_date + relativedelta(days=-1))
+            if start_date is not None
+            else Money(0)
+        )
+
+        for transaction in self.transactions_range(start_date, end_date):
+            balance += transaction.amount
+            yield RunningBalance(transaction, balance)
+
     def balance(self, as_of: date | None = None):
         """Calculate the account balance as of the given date."""
         as_of = as_of or date.today()
@@ -71,7 +92,7 @@ class Account:
 
     def daily_balance(
         self, start_date: date | None = None, end_date: date | None = None
-    ) -> Generator[Balance]:
+    ) -> Generator[DailyBalance]:
         """
         Iterate over the daily balance of the account, yielding tuples of date
         and balance.
@@ -85,15 +106,17 @@ class Account:
         end_date = end_date or date.today()
 
         balance = self.balance(start_date)
-        yield Balance(start_date, balance)
+        yield DailyBalance(start_date, balance)
 
         for _date, delta in self._daily_balance_delta(
             start_date + timedelta(days=1), end_date
         ):
             balance += delta
-            yield Balance(_date, balance)
+            yield DailyBalance(_date, balance)
 
-    def _deltas_by_date(self, start_date: date, end_date: date) -> Generator[Balance]:
+    def _deltas_by_date(
+        self, start_date: date, end_date: date
+    ) -> Generator[DailyBalance]:
         """
         Iterate over the deltas in the account balance for each date in the
         given range, including the given start and end dates.
@@ -102,7 +125,7 @@ class Account:
         element is the total amount of all transactions on that date.
         """
         yield from (
-            Balance(
+            DailyBalance(
                 _date, Money.sum(transaction.amount for transaction in transactions)
             )
             for _date, transactions in groupby(
@@ -112,7 +135,7 @@ class Account:
 
     def _daily_balance_delta(
         self, start_date: date, end_date: date
-    ) -> Generator[Balance]:
+    ) -> Generator[DailyBalance]:
         """
         Calculate the daily change in account balance over the specified date range.
 
@@ -122,12 +145,12 @@ class Account:
         transactions to ensure a continuous range.
         """
         yield from (
-            Balance(_date, Money.sum(delta.balance for delta in deltas))
+            DailyBalance(_date, Money.sum(delta.balance for delta in deltas))
             for _date, deltas in groupby(
                 merge(
                     self._deltas_by_date(start_date, end_date),
                     (
-                        Balance(_date, Money(0))
+                        DailyBalance(_date, Money(0))
                         for _date in daterange(start_date, end_date)
                     ),
                 ),